Improve page load by moving hidden navigation div the bottom of the page.
The Navigation <div> was one of the first things to load in browser, yet
it's hidden on page load. This means a larger "avatar" image would be
one of the first elements that the browser spends times loading, even
though it may never be viewed.

After making this change, the browser renders the page just the same,
but I saw total load times on my own blog drop from about 3 seconds to
2.5 seconds-- A notable improvement.